+++ core/Fractal.java
@@ -247,7 +247,14 @@ public class Fractal {
                 }
             }
         }
+    }
 
+    public void setJuliaSeed(double reSeed, double imSeed) {
+        if (rendering) return;
+
+        juliaReSeed = reSeed;
+        juliaImSeed = imSeed;
+        generate();
     }
 
     /**
@@ -335,7 +342,7 @@ public class Fractal {
 
         long endTime = System.currentTimeMillis();
         long duration = (endTime - startTime);
-        System.out.println("Rendered in: " + duration + "ms");
+//        System.out.println("Rendered in: " + duration + "ms");
     }
 
     /**